Chore(Folder Structure): Chage folder structure to improve usability
