+++ source/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['preProcess', 'model','train']
+__all__ = ['image','model','utils','train']
 
 from os import path
 import sys
@@ -8,8 +8,8 @@ root = path.abspath('..')
 if root not in sys.path:
     sys.path.append(root)
 
+from source import image
 from source import model
-from source import preProcess
 from source import utils
 
 from source.train import train
