--- source/preProcess/__init__.py
@@ -1,16 +1,0 @@
-__all__ = ['utils']
-
-from os import path
-import sys
-
-root = path.abspath('../..')
-
-if root not in sys.path:
-    sys.path.append(root)
-
-from source.preProcess import utils
-from source.preProcess.utils import (
-    make_dir,
-    image_to_array,
-    save_array
-)
