+++ source/utils/__init__.py
@@ -9,4 +9,7 @@ if root not in sys.path:
     sys.path.append(root)
 
 from source.utils import structure
-from source.utils.structure import exists
+from source.utils.structure import exists, make_dir
+
+from source.utils import utils
+from source.utils.utils import image_to_array, save_array
